Correct a bug in detecting unexpected end.
Previously, the two following ways of parsing content would yield
different error messages, even though the content is an empty string
both times.

  $ echo -n | jiffy
  Unexpected end of input

  $ echo -n > empty && jiffy empty
  Unexpected token at position 0

The reason for the differing behavior was that STDIN does not close
after reaching end-of-file, but ARGF does. With this patch, both ways
will return unexpected end of input.